+++ apps/examples/src/examples/GPTExample/aistuff.ts
